pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "behold-urwar-droid-concepts"
version = "0.1.0"
description = "A web interface for training and testing duck droid models"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "flask>=2.0.1",
    "paramiko>=2.8.1",
    "onnxruntime>=1.10.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.3",
    "torch>=1.8.1",
    "pyyaml>=5.4.1",
    "uv>=0.1.0",
    "python-dotenv>=0.19.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
    "app/**/*.py",
    "templates/**/*.html",
    "static/**/*",
    "docs/**/*.md",
]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 